+++ openeogeotrellis/stac_save_result.py
@@ -1,4 +1,5 @@
 import json
+import logging
 import os
 import shutil
 from pathlib import Path
@@ -10,103 +11,7 @@ import requests
 from flask import Response
 from openeo_driver.datastructs import StacAsset
 from openeo_driver.save_result import SaveResult
-
-
-def get_files_from_stac_catalog(catalog_path: Union[str, Path]) -> list:
-    """
-    Goes through the stac catalog recursively to find all files.
-    """
-    if isinstance(catalog_path, str) and catalog_path.startswith("http"):
-        response = requests.get(catalog_path)
-        response.raise_for_status()
-        catalog_json = response.json()
-    else:
-        catalog_path = str(catalog_path)
-        assert os.path.exists(catalog_path)
-        catalog_json = json.loads(Path(catalog_path).read_text())
-
-    all_files = []
-    links = []
-    if "links" in catalog_json:
-        links.extend(catalog_json["links"])
-    if "assets" in catalog_json:
-        links.extend(list(catalog_json["assets"].values()))
-    for link in links:
-        if "href" in link:
-            href = link["href"]
-            if href.startswith("file://"):
-                href = href[7:]
-            href = urljoin(catalog_path, href)
-            all_files.append(href)
-
-            if "rel" in link and (link["rel"] == "child" or link["rel"] == "item"):
-                all_files.extend(get_files_from_stac_catalog(href))
-    return all_files
-
-
-def get_assets_from_stac_catalog(catalog_path: Union[str, Path]) -> Dict[str, StacAsset]:
-    if isinstance(catalog_path, str) and catalog_path.startswith("http"):
-        response = requests.get(catalog_path)
-        response.raise_for_status()
-        catalog_json = response.json()
-    else:
-        catalog_path = str(catalog_path)
-        assert os.path.exists(catalog_path)
-        catalog_json = json.loads(Path(catalog_path).read_text())
-
-    all_assets = {}
-    links = []
-    if "links" in catalog_json:
-        links.extend(catalog_json["links"])
-    if "assets" in catalog_json:
-        links.extend(list(catalog_json["assets"].values()))
-        assets = catalog_json["assets"]
-        all_assets.update(assets)
-    for link in links:
-        if "href" in link:
-            href = link["href"]
-            if href.startswith("file://"):
-                href = href[7:]
-            href = urljoin(catalog_path, href)
-
-            if "rel" in link and (link["rel"] == "child" or link["rel"] == "item"):
-                all_assets.update(get_assets_from_stac_catalog(href))
-    return all_assets
-
-
-def get_items_from_stac_catalog(catalog_path: Union[str, Path], make_hrefs_absolute=False) -> dict:
-    if isinstance(catalog_path, str) and catalog_path.startswith("http"):
-        response = requests.get(catalog_path)
-        response.raise_for_status()
-        catalog_json = response.json()
-    else:
-        catalog_path = str(catalog_path)
-        assert os.path.exists(catalog_path), f"catalog_path does not exist: {catalog_path}"
-        catalog_json = json.loads(Path(catalog_path).read_text())
-
-    all_items = {}
-    links = []
-    if "links" in catalog_json:
-        links.extend(catalog_json["links"])
-    if "assets" in catalog_json:
-        links.extend(list(catalog_json["assets"].values()))
-        all_items.update({catalog_json["id"]: catalog_json})
-    if make_hrefs_absolute:
-        for item in all_items.values():
-            if "assets" in item:
-                for asset in item["assets"].values():
-                    if "href" in asset:
-                        asset["href"] = urljoin(catalog_path, asset["href"])
-    for link in links:
-        if "href" in link:
-            href = link["href"]
-            if href.startswith("file://"):
-                href = href[7:]
-            href = urljoin(catalog_path, href)
-
-            if "rel" in link and (link["rel"] == "child" or link["rel"] == "item"):
-                all_items.update(get_items_from_stac_catalog(href, make_hrefs_absolute))
-    return all_items
+from openeo_driver.util.stac_utils import get_files_from_stac_catalog, get_items_from_stac_catalog
 
 
 class StacSaveResult(SaveResult):
@@ -128,7 +33,7 @@ class StacSaveResult(SaveResult):
         Copy over the stac catalog to the expected directory
         :return: STAC assets dictionary: https://github.com/radiantearth/stac-spec/blob/master/item-spec/item-spec.md#assets
         """
-        stac_assets = get_files_from_stac_catalog(self.stac_root)
+        stac_assets = get_files_from_stac_catalog(self.stac_root, include_metadata=True)
         if str(directory).endswith("out"):
             directory = str(directory)[:-4]
 
@@ -147,6 +52,7 @@ class StacSaveResult(SaveResult):
             dest_path = urljoin(str(directory) + "/", relative_path)
             Path(dest_path).parent.mkdir(parents=True, exist_ok=True)
             if asset_path_parsed.scheme in ("http", "https"):
+                logging.info(f"URL: copy_asset({asset_path})")
                 response = requests.get(asset_path)
                 response.raise_for_status()
                 with open(dest_path, "wb") as f: