+++ tests/test_stac_save_result.py
@@ -5,49 +5,10 @@ from pystac import Collection
 
 from openeogeotrellis.stac_save_result import (
     StacSaveResult,
-    get_files_from_stac_catalog,
-    get_assets_from_stac_catalog,
-    get_items_from_stac_catalog,
 )
 
 repository_root = Path(__file__).parent.parent
 print(repository_root)
-
-
-def test_get_files_from_stac_catalog_path():
-    stac_root = repository_root / "docker/local_batch_job/example_stac_catalog/collection.json"
-    ret = get_files_from_stac_catalog(stac_root)
-    print(ret)
-    assert len(ret) == 6
-
-
-def test_get_files_from_stac_catalog_url():
-    stac_root = "https://raw.githubusercontent.com/Open-EO/openeo-geopyspark-driver/refs/heads/master/docker/local_batch_job/example_stac_catalog/collection.json"
-    ret = get_files_from_stac_catalog(stac_root)
-
-    print(ret)
-    assert len(ret) == 6
-
-
-def test_get_assets_from_stac_catalog():
-    stac_root = repository_root / "docker/local_batch_job/example_stac_catalog/collection.json"
-    ret = get_assets_from_stac_catalog(stac_root)
-    print(ret)
-    assert len(ret.values()) == 3
-
-
-def test_get_items_from_stac_catalog():
-    stac_root = repository_root / "docker/local_batch_job/example_stac_catalog/collection.json"
-    ret = get_items_from_stac_catalog(stac_root)
-    print(ret)
-    assert len(ret) == 3
-
-
-def test_get_items_from_stac_catalog_recursive():
-    stac_root = str(repository_root / "tests/data/stac/recursive-stac-example/collection.json")
-    ret = get_items_from_stac_catalog(stac_root, make_hrefs_absolute=True)
-    print(ret)
-    assert len(ret) == 3
 
 
 def test_stac_save_result():
@@ -61,7 +22,12 @@ def test_stac_save_result():
     sr = StacSaveResult(stac_root)
     ret = sr.write_assets(tmp_dir)
     Collection.from_file(sr.stac_root_local).validate_all()
+
     print(ret)
+    for key, item in ret.items():
+        for asset in item["assets"].values():
+            assert Path(asset["href"]).exists()
+
     assert len(ret) == 3